Correct units & calculation
A limitation in Python prevents `''` from being iterated over
properly, which is pretty much yet another thing in a long list of
things that I don't like about Python and that Perl does better!
However, it is easily manageable by tweaking the logic, at least in
this case.

Also, made it start with bytes, not with Ks.